+++ lib/libtpe/tpe/task/eth_block_compression_task.h
@@ -29,7 +29,7 @@ public:
         PBlockBuffer_t block_buffer,
         PTransactionService_t tx_service,
         bool enable_block_compression,
-        float min_tx_age_seconds
+        double min_tx_age_seconds
     );
 
     PByteArray_t bx_block();
@@ -61,7 +61,7 @@ private:
     size_t _content_size;
     size_t _starting_offset;
     bool _enable_block_compression;
-    float _min_tx_age_seconds;
+    double _min_tx_age_seconds;
 };
 
 } // task
