+++ lib/libtpe/tpe/task/sub_task/btc_block_compression_sub_task.cpp
@@ -28,7 +28,7 @@ void BtcBlockCompressionSubTask::init(
     const BlockBuffer_t* block_buffer,
     POffests_t tx_offsets,
     bool enable_block_compression,
-    float min_tx_age_seconds
+    double min_tx_age_seconds
 )
 {
 	_tx_service = tx_service;
@@ -57,7 +57,7 @@ const std::vector<unsigned int>& BtcBlockCompressionSubTask::ignored_short_ids()
 
 void BtcBlockCompressionSubTask::_execute()  {
 	size_t output_offset = 0;
-    float max_timestamp_for_compression = std::chrono::duration_cast<std::chrono::seconds>(
+    double max_timestamp_for_compression = std::chrono::duration_cast<std::chrono::seconds>(
         std::chrono::system_clock::now().time_since_epoch()
     ).count() - _min_tx_age_seconds;
 
@@ -73,7 +73,7 @@ void BtcBlockCompressionSubTask::_execute()  {
 						offset
 		));
 
-		float short_id_assign_time = 0.0;
+		double short_id_assign_time = 0.0;
 		if (_tx_service->has_short_id(sha)) {
             short_id_assign_time = _tx_service->get_short_id_assign_time(_tx_service->get_short_id(sha));
 		}
