+++ lib/libtpe/tpe/task/sub_task/ont_block_compression_sub_task.h
@@ -26,7 +26,7 @@ public:
         const BlockBuffer_t* block_buffer,
         POffests_t tx_offsets,
         bool enable_block_compression,
-        float min_tx_age_seconds
+        double min_tx_age_seconds
     );
 
     const utils::common::ByteArray& output_buffer();
@@ -45,7 +45,7 @@ private:
     std::vector<unsigned int> _short_ids;
     std::vector<unsigned int> _ignored_short_ids;
     bool _enable_block_compression;
-    float _min_tx_age_seconds;
+    double _min_tx_age_seconds;
 };
 
 }
